Create MyGraph
Method rndfill(): Fills the adjacency matrix with random values 0 and 1. The graph becomes directed and connected.

Method findpath(): Implements a search for an Euler path in a graph and outputs a possible path. The Fleury algorithm is used. The method uses a stack to track the path and a `B` matrix to mark the edges traversed.

Method print(): Prints the graph's adjacency matrix to the screen.

Method manfill(): Manually filling the adjacency matrix. The graph becomes undirected and connected.

Method getsize(): Returns the size of the graph.

Method setsize(int size): Sets the new size of the graph.